Check interactivity before clearing terminal
If running `tide configure --auto` as part of some larger non-interactive
process i.e. in the context of some .dotfiles installer then clearing the
screen is quite jarring as it removes the logs of the parent process.

The patch adds an interactivity check to keep the current behaviour when
running `tide configure` in a tty, but makes it a better behaved citizen
in automation contexts.